requirements.txt
mcp>=1.0.0
twikit>=1.7.0
pydantic>=2.0.0
python-dotenv>=1.0.0
